+++ include/quill/backend/ThreadUtilities.h
@@ -12,7 +12,6 @@
 #include <cstdint>
 #include <cstring>
 #include <string>
-#include <vector>
 
 #if defined(_WIN32)
   #if !defined(WIN32_LEAN_AND_MEAN)
@@ -28,7 +27,6 @@
 
   #include <codecvt>
   #include <locale>
-  #include <tlhelp32.h>
 #elif defined(__APPLE__)
   #include <mach/thread_act.h>
   #include <mach/thread_policy.h>
@@ -214,67 +212,6 @@ QUILL_NODISCARD QUILL_EXPORT QUILL_ATTRIBUTE_USED inline uint32_t get_thread_id(
 #endif
 }
 
-/**
- * Returns the names of all threads in the process
- * @return the names of all threads
- */
-QUILL_NODISCARD inline std::vector<std::string> get_current_process_thread_names()
-{
-  std::vector<std::string> thread_names;
-
-#if defined(__CYGWIN__) || defined(__MINGW32__) || defined(__MINGW64__) || defined(QUILL_NO_THREAD_NAME_SUPPORT)
-  // Disabled on MINGW / Cygwin.
-#elif defined(_WIN32)
-  DWORD current_process_id = GetCurrentProcessId();
-  HANDLE snapshot_handle = CreateToolhelp32Snapshot(TH32CS_SNAPTHREAD, 0);
-
-  if (snapshot_handle == INVALID_HANDLE_VALUE)
-  {
-    QUILL_THROW(QuillError{"Failed to create thread snapshot."});
-  }
-
-  THREADENTRY32 te = {sizeof(te)};
-  for (BOOL ok = Thread32First(snapshot_handle, &te); ok; ok = Thread32Next(snapshot_handle, &te))
-  {
-    if (te.th32OwnerProcessID == current_process_id)
-    {
-      HANDLE thread_handle = OpenThread(THREAD_QUERY_LIMITED_INFORMATION, FALSE, te.th32ThreadID);
-      if (thread_handle)
-      {
-        PWSTR thread_name = nullptr;
-
-        typedef HRESULT(WINAPI * GetThreadDescriptionSignature)(HANDLE, PWSTR*);
-        HRESULT hr = callRunTimeDynamicLinkedFunction<HRESULT, GetThreadDescriptionSignature>(
-          "KernelBase.dll", "GetThreadDescription", thread_handle, &thread_name);
-
-        if (FAILED(hr))
-        {
-          CloseHandle(thread_handle);
-          QUILL_THROW(QuillError{"Failed to get thread name"});
-        }
-
-        if (QUILL_UNLIKELY(thread_name == nullptr))
-        {
-          CloseHandle(thread_handle);
-          QUILL_THROW(QuillError{"Failed to get thread name. Invalid data."});
-        }
-
-        std::wstring const wide_name{thread_name, wcsnlen_s(thread_name, 256)};
-        LocalFree(thread_name);
-        CloseHandle(thread_handle);
-
-        thread_names.push_back(ws2s(wide_name));
-      }
-    }
-  }
-
-  CloseHandle(snapshot_handle);
-#else
-  // Apple, linux
-#endif
-
-  return thread_names;
-}
 } // namespace detail
 
 QUILL_END_NAMESPACE